pyproject.toml
[project]
name = "codelens"
version = "0.1.0"
description = "Feature Intelligence Explorer - understand feature history from Jira, Slack, and code"
requires-python = ">=3.11"
dependencies = [
    "cognee>=0.1.19",
    "fastapi>=0.115.0",
    "llama-cpp-python>=0.3.2",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.12.0",
    "requests>=2.31.0",
    "uvicorn>=0.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared", "pipeline"]
